Fixed wrong filename bug with specialchars
When adding a folder that contains '&' (and probably other special chars):
os.path.dirname sticks to '&', but gnomevfs.URI converts this to "%26".

The filelist is created with gnomevfs.URI(uri) --> "%26"
But for 'base', only os.path.dirname is used --> "&"
In SoundFile.__init__(), self.filename is created by cutting off the
length of the base. This lead to wrong filenames, so the output of
soundconverter was stored in wrong, newly created directories.